// src/app/State/todo.types.ts
/* eslint-disable @typescript-eslint/naming-convention */
export enum TodoActionTypes{
  GET_TODO = '[GET_TODO]',
  GET_TODO_SUCCESS = '[GET_TODO] SUCCESS',
  GET_TODO_FAILURE = '[GET_TODO_FAILURE] FAILURE',

  ADD_TODO = '[ADD_TODO]',
  ADD_TODO_SUCCESS = '[ADD_TODO] SUCCESS',
  ADD_TODO_FAILURE = '[ADD_TODO_FAILURE] FAILURE',

  DELETE_TODO = '[DELETE_TODO]',
  DELETE_TODO_SUCCESS = '[DELETE_TODO] SUCCESS',
  DELETE_TODO_FAILURE = '[DELETE_TODO_FAILURE] FAILURE',

  UPDATE_TODO = '[UPDATE_TODO]',
  UPDATE_TODO_SUCCESS = '[UPDATE_TODO] SUCCESS',
  UPDATE_TODO_FAILURE = '[UPDATE_TODO_FAILURE] FAILURE',
}